refactor challenge form for reusability
