+++ src/app/dashboard/challenges/create/page.tsx
@@ -1,10 +1,16 @@
+"use client";
+
 import Heading from "@/components/heading";
-import CreateChallengeForm from "./create-challenge-form";
 import Link from "next/link";
 import { ChevronLeft } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import ChallengeForm from "../challenge-form";
+import { useMutation } from "convex/react";
+import { api } from "@/api";
 
 export default function CreateChallengePage() {
+  const createChallenge = useMutation(api.challenges.createChallenge);
+
   return (
     <>
       <div className="mb-8 flex items-center gap-4">
@@ -15,7 +21,7 @@ export default function CreateChallengePage() {
         </Button>
         <Heading className="mb-0">Create Challenge</Heading>
       </div>
-      <CreateChallengeForm />
+      <ChallengeForm mutation={createChallenge} />
     </>
   );
 }
